requirements.txt
colorama==0.4.3
google_api_python_client==2.107.0
httplib2==0.22.0
protobuf==4.25.1
Requests==2.31.0
setuptools==69.0.3